rm zarr.codecs.mixins
