+++ src/zarr/codecs/gzip.py
@@ -4,7 +4,7 @@ from dataclasses import dataclass
 from typing import TYPE_CHECKING
 
 from numcodecs.gzip import GZip
-from zarr.codecs.mixins import BytesBytesCodecBatchMixin
+from zarr.abc.codec import BytesBytesCodec
 from zarr.buffer import Buffer, as_numpy_array_wrapper
 from zarr.codecs.registry import register_codec
 from zarr.common import parse_named_configuration, to_thread
@@ -26,7 +26,7 @@ def parse_gzip_level(data: JSON) -> int:
 
 
 @dataclass(frozen=True)
-class GzipCodec(BytesBytesCodecBatchMixin):
+class GzipCodec(BytesBytesCodec):
     is_fixed_size = False
 
     level: int = 5
@@ -44,14 +44,14 @@ class GzipCodec(BytesBytesCodecBatchMixin):
     def to_dict(self) -> Dict[str, JSON]:
         return {"name": "gzip", "configuration": {"level": self.level}}
 
-    async def decode_single(
+    async def _decode_single(
         self,
         chunk_bytes: Buffer,
         _chunk_spec: ArraySpec,
     ) -> Buffer:
         return await to_thread(as_numpy_array_wrapper, GZip(self.level).decode, chunk_bytes)
 
-    async def encode_single(
+    async def _encode_single(
         self,
         chunk_bytes: Buffer,
         _chunk_spec: ArraySpec,
