+++ src/zarr/codecs/sharding.py
@@ -6,14 +6,17 @@ from functools import lru_cache
 
 
 import numpy as np
-from zarr.abc.codec import ByteGetter, ByteSetter, Codec, CodecPipeline
+from zarr.abc.codec import (
+    ArrayBytesCodec,
+    ArrayBytesCodecPartialDecodeMixin,
+    ArrayBytesCodecPartialEncodeMixin,
+    ByteGetter,
+    ByteSetter,
+    Codec,
+    CodecPipeline,
+)
 from zarr.codecs.bytes import BytesCodec
 from zarr.codecs.crc32c_ import Crc32cCodec
-from zarr.codecs.mixins import (
-    ArrayBytesCodecBatchMixin,
-    ArrayBytesCodecPartialDecodeBatchMixin,
-    ArrayBytesCodecPartialEncodeBatchMixin,
-)
 from zarr.codecs.pipeline import BatchedCodecPipeline
 from zarr.codecs.registry import register_codec
 from zarr.common import (
@@ -286,9 +289,7 @@ class _MergingShardBuilder(ShardMutableMapping):
 
 @dataclass(frozen=True)
 class ShardingCodec(
-    ArrayBytesCodecBatchMixin,
-    ArrayBytesCodecPartialDecodeBatchMixin,
-    ArrayBytesCodecPartialEncodeBatchMixin,
+    ArrayBytesCodec, ArrayBytesCodecPartialDecodeMixin, ArrayBytesCodecPartialEncodeMixin
 ):
     chunk_shape: ChunkCoords
     codecs: CodecPipeline
@@ -373,7 +374,7 @@ class ShardingCodec(
                 + "shard's inner `chunk_shape`."
             )
 
-    async def decode_single(
+    async def _decode_single(
         self,
         shard_bytes: Buffer,
         shard_spec: ArraySpec,
@@ -415,7 +416,7 @@ class ShardingCodec(
 
         return out
 
-    async def decode_partial_single(
+    async def _decode_partial_single(
         self,
         byte_getter: ByteGetter,
         selection: SliceSelection,
@@ -476,7 +477,7 @@ class ShardingCodec(
         )
         return out
 
-    async def encode_single(
+    async def _encode_single(
         self,
         shard_array: NDBuffer,
         shard_spec: ArraySpec,
@@ -511,7 +512,7 @@ class ShardingCodec(
 
         return await shard_builder.finalize(self.index_location, self._encode_shard_index)
 
-    async def encode_partial_single(
+    async def _encode_partial_single(
         self,
         byte_setter: ByteSetter,
         shard_array: NDBuffer,
